Make resttemaplte autowired
