+++ src/test/java/uk/gov/hmcts/reform/sandl/snlapi/controllers/GetWelcomeTest.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.sandl.snlapi.controllers;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.client.AutoConfigureWebClient;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
@@ -14,6 +15,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @RunWith(SpringRunner.class)
 @WebMvcTest(RootController.class)
+@AutoConfigureWebClient
 public class GetWelcomeTest {
 
     @Autowired
